requirements.txt
attrs>=22.2.0
certifi==2022.12.7
cffi>=1.17.0
cfgv==3.3.1
charset-normalizer==2.0.7
cryptography>=46.0.0
distlib==0.3.6
exceptiongroup==1.1.1
filelock>=3.13
identify==2.5.23
idna==3.3
iniconfig==1.1.1
isort==5.12.0
nodeenv==1.7.0
numpy>=1.24.3
oauthlib==3.1.1
packaging>=22.0
pandas>=2.0.3
platformdirs==3.4.0
pluggy==1.0.0
pre-commit==3.2.2
py==1.11.0
pycparser==2.21
pyotp==2.6.0
pyparsing==2.4.7
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv>=0.21.0
pytz>=2022.5
PyYAML>=6.0
requests>=2.27.1
requests-oauthlib==1.3.0
robin-stocks==2.1.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
tweepy==4.3.0
tzdata==2023.3
urllib3>=1.26.11
vaderSentiment==3.3.2
virtualenv==20.22.0
yfinance==0.2.49
textblob==0.17.1